% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyttest_pairwise.R
\name{svyttest_pairwise}
\alias{svyttest_pairwise}
\title{Design-based pairwise two-sample t-test}
\usage{
svyttest_pairwise(formula, by, design, df.resid = survey::degf(design) - 2)
}
\arguments{
\item{formula}{One-sided formula indicating the outcome variable(s)}

\item{by}{One-sided formula indicating the by variable(s) - these are done jointly if more than one}

\item{design}{survey design object, could be subset}

\item{df.resid}{degrees of freedom for testing, defaults to degf(design)-2}
}
\value{
a data.frame with the label of the comparison, the contrast estimate, the se of the contrast estimate, the t-value, the degrees of freedom, and the p-value
}
\description{
Calculate pairwise two-sample t-tests. This function is a wrapper for
svyby() and svycontrast() from the survey package
}
\examples{
data("api", package = "survey")
dclus1 <- survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)
svyttest_pairwise(~api00, ~stype, dclus1)
svyttest_pairwise(~api00, ~stype, dclus1, df.resid = 10)
svyttest_pairwise(~api00, ~ stype + comp.imp, dclus1)
svyttest_pairwise(~ I(api00 - api99), ~ stype + comp.imp, dclus1)
\dontrun{
svyttest_pairwise(~ api00 + api99, ~ stype + comp.imp, dclus1)
}
}
